requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
passlib[bcrypt]
python-jose
email-validator
cryptography
google-generativeai